refactor: int -> Long 객체 변경

+++ src/main/kotlin/part/soten/user/UserResponse.kt
@@ -3,4 +3,4 @@ package part.soten.user
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserResponse(val id: Int = 5, val name: String, val email: String)
+data class UserResponse(val id: Long = 5, val name: String, val email: String)
